collision changes

+++ Dane/Circle.cs
@@ -4,14 +4,14 @@ namespace Dane
 {
     internal class Circle
     {
-        private int x;
-        private int y;
+        private double x;
+        private double y;
         private int radius;
-        private int vx;
-        private int vy;
+        private double vx;
+        private double vy;
         private int mass;
 
-        public Circle(int x, int y, int radius, int vx, int vy, int mass)
+        public Circle(double x, double y, int radius, double vx, double vy, int mass)
         {
             this.x = x; this.y = y;
             this.radius = radius;
@@ -21,16 +21,16 @@ namespace Dane
 
         public Circle() : this(0,0,10, 5, 5, 10) { }
 
-        public int GetX() { return x; }
-        public int GetY() { return y; }
+        public double GetX() { return x; }
+        public double GetY() { return y; }
         public int GetRadius() { return radius; }
-        public int GetVx() { return vx; }
-        public int GetVy() { return vy; }
+        public double GetVx() { return vx; }
+        public double GetVy() { return vy; }
         public int GetMass() { return mass; }
 
-        public void SetX(int x) { this.x = x; }
-        public void SetY(int y) { this.y = y; }
-        public void SetVx(int vx) { this.vx = vx; }
-        public void SetVy(int vy) { this.vy = vy; }
+        public void SetX(double x) { this.x = x; }
+        public void SetY(double y) { this.y = y; }
+        public void SetVx(double vx) { this.vx = vx; }
+        public void SetVy(double vy) { this.vy = vy; }
     }
 }
